+++ src/Sophia_Script_for_Windows_11/Module/Sophia.psm1
@@ -2,8 +2,8 @@
 	.SYNOPSIS
 	Sophia Script is a PowerShell module for Windows 10 & Windows 11 fine-tuning and automating the routine tasks
 
-	Version: v6.4.1
-	Date: 14.03.2023
+	Version: v6.4.2
+	Date: 20.03.2023
 
 	Copyright (c) 2014—2023 farag
 	Copyright (c) 2019—2023 farag & Inestic
@@ -7862,7 +7862,7 @@ function ThumbnailCacheRemoval
 	{
 		"Disable"
 		{
-			New-ItemProperty -Path "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\VolumeCaches\Thumbnail Cache" -Name Autorun -PropertyType DWord -Value 0 -Force ###
+			New-ItemProperty -Path "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\VolumeCaches\Thumbnail Cache" -Name Autorun -PropertyType DWord -Value 0 -Force
 			New-ItemProperty -Path "HKLM:\SOFTWARE\WOW6432Node\Microsoft\Windows\CurrentVersion\Explorer\VolumeCaches\Thumbnail Cache" -Name Autorun -PropertyType DWord -Value 0 -Force
 		}
 		"Enable"
@@ -12974,7 +12974,7 @@ function DNSoverHTTPS
 	.NOTES
 	Machine-wide
 #>
-function LocalSecurityAuthority ###
+function LocalSecurityAuthority
 {
 	param
 	(