+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 5.16.2 | 6.4.2 — 20.03.2023
+
+* Added `LocalSecurityAuthority` function to prevent code injection (for Windows 11 22H2 only);
+  * <https://learn.microsoft.com/en-us/windows-server/security/credentials-protection-and-management/configuring-additional-lsa-protection>
+* Improved `Checks` having expanded the list of blocked harmful tweakers;
+* Fixed `ThumbnailCacheRemoval` function;
+* Removed `SnapAssistFlyout` function;
+* Minor changes.
+
+### Wrapper 2.6.4
+
+* Minor changes.
+
 ## 5.16.1 | 6.4.1 — 14.03.2023
 
 * Code refactoring;